+++ src/app/about/components/AboutBanner.tsx
@@ -1,0 +1,45 @@
+import { Box } from '@mui/material';
+import { styled } from '@mui/material/styles';
+
+export const AboutHeader = styled(Box)(({ theme }) => ({
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    backgroundColor: theme.palette.background.dark,
+    padding: theme.spacing(10, 50),
+    color: theme.palette.text.primary,
+    gap: theme.spacing(5),
+    [theme.breakpoints.down('lg')]: {
+        padding: theme.spacing(8, 20),
+    },
+    [theme.breakpoints.down('md')]: {
+        padding: theme.spacing(8, 20),
+    },
+    [theme.breakpoints.down('sm')]: {
+        padding: theme.spacing(5, 5),
+        flexDirection: 'column',
+        textAlign: 'center',
+    },
+    [theme.breakpoints.only('sm')]: {
+        padding: theme.spacing(5, 5),
+        flexDirection: 'column',
+        textAlign: 'center',
+    },
+}));
+
+export const HeaderImage = styled('img')(({ theme }) => ({
+    borderRadius: theme.shape.borderRadius,
+    width: '100%',
+    maxWidth: '576px',
+    height: 'auto',
+    objectFit: 'cover',
+}));
+
+export const HeaderTitle = styled('h1')(({ theme }) => ({
+    color: 'white',
+    fontSize: theme.typography.h1.fontSize,
+    fontFamily: theme.typography.h1.fontFamily,
+    fontWeight: theme.typography.h1.fontWeight,
+    margin: 0,
+    lineHeight: 1.2,
+}));
