--- src/app/about/components/BodyText.tsx
@@ -1,9 +1,0 @@
-import { styled } from '@mui/material/styles';
-
-export const BodyText = styled('p')(({ theme }) => ({
-    color: 'white',
-    marginTop: theme.spacing(2),
-    marginBottom: 0,
-    fontSize: theme.typography.body1.fontSize,
-    fontFamily: theme.typography.fontFamily,
-}));
