--- src/app/about/components/HeaderImage.tsx
@@ -1,9 +1,0 @@
-import { styled } from '@mui/material/styles';
-
-export const HeaderImage = styled('img')(({ theme }) => ({
-    borderRadius: theme.shape.borderRadius,
-    width: '100%',
-    maxWidth: '576px',
-    height: 'auto',
-    objectFit: 'cover'
-}));
