+++ src/app/about/layout.tsx
@@ -1,0 +1,13 @@
+import AboutUsLayoutStyled from '@/components/layout/aboutus-layout';
+
+export default function AboutUsLayout({
+    children,
+}: Readonly<{
+    children: React.ReactNode;
+}>) {
+    return (
+        <AboutUsLayoutStyled>
+            {children}
+        </AboutUsLayoutStyled>
+    );
+}
