+++ src/components/layout/aboutus-layout/index.tsx
@@ -1,0 +1,26 @@
+import { Container } from '@mui/material';
+
+import Footer from "../main-layout/Footer"
+import Navbar from "../main-layout/Navbar"
+
+export default function AboutUsLayout({
+    children,
+}: Readonly<{
+    children: React.ReactNode;
+}>) {
+    return (
+        <>
+            <Container maxWidth="xl">
+            <Navbar variant='dark'/>
+            </Container>
+            
+            <Container maxWidth="xl">
+                {children}
+            </Container>
+            
+            <Container maxWidth="xl">
+                <Footer />
+            </Container>
+        </>
+    );
+}
